# src/main/resources/openapi.yml
openapi: 3.0.0
info:
  title: Cruise API
  description: API for retrieving cruise details
  version: 1.0.0
paths:
  /cruise-details:
    get:
      summary: Get Cruise Details
      tags:
        - n-c-controller
      operationId: getCruiseDetails
      description: Returns the details of a specific cruise
      responses:
        '204':
          description: Successfully retrieved the cruise details
        '400':
          description: Bad Request - Invalid request parameter
        '404':
          description: Not Found - Cruise not found
  /cruise:
    get:
      summary: Retrieve Cruises by Departure Date and Destination Code
      tags:
        - n-c-controller
      operationId: getSailingsByDestinationAndDeparture
      parameters:
        - $ref: '#/components/parameters/departureDate'
        - $ref: '#/components/parameters/destinationCodeRequired'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CruiseOverView'
  /best-cruise-price:
    get:
      summary: Return the Best Cruise Price for a Sailing ID
      tags:
        - n-c-controller
      operationId: getBestSailingByPriceAndType
      parameters:
        - $ref: '#/components/parameters/sailId'
        - $ref: '#/components/parameters/roomType'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sailings'
  /daily-price-drops:
    get:
      summary: Return Daily Price Drops
      tags:
        - n-c-controller
      operationId: getSailingsPriceDrops
      parameters:
        - $ref: '#/components/parameters/fromDate'
        - $ref: '#/components/parameters/toDate'
        - $ref: '#/components/parameters/percentage'
        - $ref: '#/components/parameters/roomType'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CruiseOverView'
  /find-cruise:
    get:
      summary: Return cruises
      tags:
        - n-c-controller
      operationId: findCruise
      parameters:
        - $ref: '#/components/parameters/departureDate'
        - $ref: '#/components/parameters/returnDate'
        - $ref: '#/components/parameters/priceUpTo'
        - $ref: '#/components/parameters/priceFrom'
        - $ref: '#/components/parameters/daysAtSeaMin'
        - $ref: '#/components/parameters/daysAtSeaMax'
        - $ref: '#/components/parameters/destinationCode'
        - $ref: '#/components/parameters/departurePort'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CruiseOverView'

  /sailings/{sailId}/prices:
    get:
      summary: Return Prices of Sailings by Sailing ID
      tags:
        - n-c-controller
      operationId: getSailingsPricesBySailId
      parameters:
        - $ref: '#/components/parameters/sailIdPathParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Sailings'
  /destination-codes:
    get:
      summary: Get Unique Destination Codes
      tags:
        - n-c-controller
      operationId: getDestinationCodes
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '400':
          description: Bad Request - Invalid request parameter
        '404':
          description: Not Found - Cruise not found
  /cruises/{destinationCode}:
    get:
      summary: Get Cruises by Destination Code
      tags:
        - n-c-controller
      operationId: getCruisesByDestinationCode
      parameters:
        - $ref: '#/components/parameters/destinationCodePathParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CruiseDetails'
        '400':
          description: Bad Request - Invalid request parameter
        '404':
          description: Not Found - Cruise not found
  /cruise/{code}/prices:
    get:
      summary: Retrieve Prices for a Cruise by Code
      tags:
        - n-c-controller
      operationId: getCruisePricesByCode
      parameters:
        - $ref: '#/components/parameters/codePathParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Sailings'

components:
# -------------------------------------------------------------------------------
  parameters:
    departureDate:
      name: departureDate
      in: query
      required: true
      description: The departure date parameter (e.g., 2023-10-08)
      schema:
        type: string
        format: date

    returnDate:
      name: returnDate
      in: query
      required: true
      description: The return date parameter (e.g., 2023-10-08)
      schema:
        type: string
        format: date

    daysAtSeaMin:
      name: daysAtSeaMin
      in: query
      description: The minimum days at sea
      schema:
        type: number

    daysAtSeaMax:
      name: daysAtSeaMax
      in: query
      description: The maximum days at sea
      schema:
        type: number

    duration:
      name: duration
      in: query
      description: The duration of the cruise
      schema:
        type: number

    priceUpTo:
      name: priceUpTo
      in: query
      description: The max price of the cruise
      schema:
        type: number

    priceFrom:
      name: priceFrom
      in: query
      description: The min price of the cruise
      schema:
        type: number

    departurePort:
      name: departurePort
      in: query
      description: The destination code parameter
      schema:
        type: string

    destinationCodeRequired:
      name: destinationCode
      in: query
      required: true
      description: The destination code parameter
      schema:
        type: string

    destinationCode:
      name: destinationCode
      in: query
      description: The destination code parameter
      schema:
        type: string
    sailId:
      name: sailId
      in: query
      required: true
      description: The sailing ID
      schema:
        type: number

    roomType:
      name: roomType
      in: query
      required: false
      description: The room type (e.g., INSIDE, OCEAN_VIEW)
      schema:
        type: string
        enum:
          - INSIDE
          - OCEAN_VIEW
          - MINI_SUITE
          - STUDIO
          - SUITE
          - HAVEN
          - SPA

    fromDate:
      name: fromDate
      in: query
      description: The from date parameter (e.g., 2023-10-08, default is yesterday)
      schema:
        type: string
        format: date

    percentage:
      name: percentage
      in: query
      schema:
        type: number

    toDate:
      name: toDate
      in: query
      description: The to date parameter (e.g., 2023-10-08, default is today)
      schema:
        type: string
        format: date

    sailIdPathParam:
      name: sailId
      in: path
      required: true
      description: The sailing ID
      schema:
        type: number

    destinationCodePathParam:
      name: destinationCode
      in: path
      required: true
      description: Destination code
      schema:
        type: string

    codePathParam:
      name: code
      in: path
      required: true
      description: Cruise code
      schema:
        type: string
# -------------------------------------------------------------------------------
  schemas:
    CruiseDetailsShared:
      type: object
      properties:
        destinationCodes:
          type: array
          items:
            type: string
        embarkationPortCode:
          type: string
        shipCode:
          type: string
        duration:
          type: number
        guestCount:
          type: number
        portsOfCall:
          type: array
          items:
            type: string

    SailingsShared:
      type: object
      properties:
        packageId:
          type: number
        sailId:
          type: number
        departureDate:
          type: string
        returnDate:
          type: string
        publishedDate:
          type: string
          format: date
        studio:
          type: number
        inside:
          type: number
        oceanView:
          type: number
        balcony:
          type: number
        miniSuite:
          type: number
        suite:
          type: number
        haven:
          type: number
        spa:
          type: number

    Pricing:
      type: object
      properties:
        code:
          type: string
        status:
          type: string
        combinedPrice:
          type: number
        totalPrice:
          type: number
        offerCodes:
          type: array
          items:
            type: string
        date:
          type: string
          format: date

    CruiseOverViewShared:
      allOf:
        - $ref: '#/components/schemas/SailingsShared'
        - $ref: '#/components/schemas/CruiseDetailsShared'

    CruiseOverView:
      allOf:
        - $ref: '#/components/schemas/CruiseOverViewShared'
      properties:
        packageId:
          type: number
        price:
          type: number

    Sailings:
      allOf:
        - $ref: '#/components/schemas/SailingsShared'
      properties:
        bundleType:
          type: string
        oldPrice:
          type: number
        pricing:
          type: array
          items:
            $ref: '#/components/schemas/Pricing'


    CruiseDetails:
      allOf:
        - $ref: '#/components/schemas/CruiseDetailsShared'
        - type: object
          properties:
            bundleType:
              type: string
            code:
              type: string
            embarkationPort:
              type: object
              properties:
                code:
                  type: string
                  description: Embarkation port code
            sailings:
              type: array
              items:
                $ref: '#/components/schemas/Sailings'